LayoutTests/platform/gtk/imported/w3c/web-platform-tests/2dcontext/imagebitmap/createImageBitmap-invalid-args-expected.txt

CONSOLE MESSAGE: line 137: Canvas area exceeds the maximum limit (width * height > 268435456).

PASS createImageBitmap with a an HTMLCanvasElement source and sw set to 0 
PASS createImageBitmap with a an HTMLCanvasElement source and sh set to 0 
FAIL createImageBitmap with a an HTMLCanvasElement source and oversized (unallocatable) crop region assert_unreached: Should have rejected: undefined Reached unreachable code
PASS createImageBitmap with a an HTMLVideoElement source and sw set to 0 
PASS createImageBitmap with a an HTMLVideoElement source and sh set to 0 
FAIL createImageBitmap with a an HTMLVideoElement source and oversized (unallocatable) crop region assert_unreached: Should have rejected: undefined Reached unreachable code
PASS createImageBitmap with a an HTMLVideoElement from a data URL source and sw set to 0 
PASS createImageBitmap with a an HTMLVideoElement from a data URL source and sh set to 0 
FAIL createImageBitmap with a an HTMLVideoElement from a data URL source and oversized (unallocatable) crop region assert_unreached: Should have rejected: undefined Reached unreachable code
PASS createImageBitmap with a a bitmap HTMLImageElement source and sw set to 0 
PASS createImageBitmap with a a bitmap HTMLImageElement source and sh set to 0 
FAIL createImageBitmap with a a bitmap HTMLImageElement source and oversized (unallocatable) crop region assert_unreached: Should have rejected: undefined Reached unreachable code
PASS createImageBitmap with a a vector HTMLImageElement source and sw set to 0 
PASS createImageBitmap with a a vector HTMLImageElement source and sh set to 0 
FAIL createImageBitmap with a a vector HTMLImageElement source and oversized (unallocatable) crop region assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL createImageBitmap with a a bitmap SVGImageElement source and sw set to 0 assert_throws: function "function () { throw e }" threw object "TypeError: Type error" ("TypeError") expected object "RangeError" ("RangeError")
FAIL createImageBitmap with a a bitmap SVGImageElement source and sh set to 0 assert_throws: function "function () { throw e }" threw object "TypeError: Type error" ("TypeError") expected object "RangeError" ("RangeError")
FAIL createImageBitmap with a a bitmap SVGImageElement source and oversized (unallocatable) crop region assert_throws: function "function () { throw e }" threw object "TypeError: Type error" ("TypeError") expected object "InvalidStateError" ("InvalidStateError")
FAIL createImageBitmap with a a vector SVGImageElement source and sw set to 0 assert_throws: function "function () { throw e }" threw object "TypeError: Type error" ("TypeError") expected object "RangeError" ("RangeError")
FAIL createImageBitmap with a a vector SVGImageElement source and sh set to 0 assert_throws: function "function () { throw e }" threw object "TypeError: Type error" ("TypeError") expected object "RangeError" ("RangeError")
FAIL createImageBitmap with a a vector SVGImageElement source and oversized (unallocatable) crop region assert_throws: function "function () { throw e }" threw object "TypeError: Type error" ("TypeError") expected object "InvalidStateError" ("InvalidStateError")
FAIL createImageBitmap with a an OffscreenCanvas source and sw set to 0 promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'testCtx.fillStyle = "rgb(255, 0, 0)"')"
FAIL createImageBitmap with a an OffscreenCanvas source and sh set to 0 promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'testCtx.fillStyle = "rgb(255, 0, 0)"')"
FAIL createImageBitmap with a an OffscreenCanvas source and oversized (unallocatable) crop region promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'testCtx.fillStyle = "rgb(255, 0, 0)"')"
PASS createImageBitmap with a an ImageData source and sw set to 0 
PASS createImageBitmap with a an ImageData source and sh set to 0 
FAIL createImageBitmap with a an ImageData source and oversized (unallocatable) crop region assert_throws: function "function () { throw e }" threw object "TypeError: createImageBitmap with ImageData is not implemented" ("TypeError") expected object "InvalidStateError" ("InvalidStateError")
PASS createImageBitmap with a an ImageBitmap source and sw set to 0 
PASS createImageBitmap with a an ImageBitmap source and sh set to 0 
FAIL createImageBitmap with a an ImageBitmap source and oversized (unallocatable) crop region assert_unreached: Should have rejected: undefined Reached unreachable code
PASS createImageBitmap with a a Blob source and sw set to 0 
PASS createImageBitmap with a a Blob source and sh set to 0 
FAIL createImageBitmap with a a Blob source and oversized (unallocatable) crop region assert_unreached: Should have rejected: undefined Reached unreachable code
PASS createImageBitmap with undefined image source. 
PASS createImageBitmap with null image source. 
PASS createImageBitmap with CanvasRenderingContext2D image source. 
PASS createImageBitmap with WebGLRenderingContext image source. 
PASS createImageBitmap with Uint8Array image source. 
PASS createImageBitmap with ArrayBuffer image source. 
PASS createImageBitmap with empty image source. 
PASS createImageBitmap with empty video source. 
PASS createImageBitmap with an oversized canvas source. 
FAIL createImageBitmap with an invalid OffscreenCanvas source. assert_throws: function "function () { throw e }" threw object "TypeError: Type error" that is not a DOMException InvalidStateError: property "code" is equal to undefined, expected 11
PASS createImageBitmap with an undecodable blob source. 
PASS createImageBitmap with a broken image source. 
FAIL createImageBitmap with an available but undecodable image source. assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL createImageBitmap with an available but zero height image source. assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL createImageBitmap with an available but zero width image source. assert_unreached: Should have rejected: undefined Reached unreachable code
PASS createImageBitmap with a closed ImageBitmap. 
